Add create delivery logic

+++ scheduler-service/src/main/java/pt/isep/tmdei/schedulerservice/model/data/CoordinatesDTO.java
@@ -1,11 +1,13 @@
 package pt.isep.tmdei.schedulerservice.model.data;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class CoordinatesDTO {
 
-    private long latitude;
-    private long longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 
 }
